chore: adds the "promptForTags" key inside "go.addTags" setting
Although the definition of "promptForTags" with the
value of false is unnecessary, it has been explicitly
placed in settings.json to facilitate the change in
situations of need, ensuring more flexibility to the
setting.

Now allows the developer to switch between the false
and true values according to the need per project.